Few fixes to run on device and removed warnings

+++ fem/dfem/doperator.hpp
@@ -48,13 +48,13 @@ public:
       prolongation_transpose(prolongation_transpose)
    {
       std::vector<Vector> s_l(solutions_l.size());
-      for (int i = 0; i < s_l.size(); i++)
+      for (size_t i = 0; i < s_l.size(); i++)
       {
          s_l[i] = *solutions_l[i];
       }
 
       std::vector<Vector> p_l(parameters_l.size());
-      for (int i = 0; i < p_l.size(); i++)
+      for (size_t i = 0; i < p_l.size(); i++)
       {
          p_l[i] = *parameters_l[i];
       }
@@ -69,7 +69,7 @@ public:
       daction_l = 0.0;
 
       prolongation(direction, direction_t, direction_l);
-      for (int i = 0; i < derivative_actions.size(); i++)
+      for (size_t i = 0; i < derivative_actions.size(); i++)
       {
          derivative_actions[i](fields_e, direction_l, daction_l);
       }
@@ -82,7 +82,7 @@ public:
       daction_l = 0.0;
 
       prolongation(transpose_direction, direction_t, direction_l);
-      for (int i = 0; i < derivative_actions_transpose.size(); i++)
+      for (size_t i = 0; i < derivative_actions_transpose.size(); i++)
       {
          derivative_actions_transpose[i](fields_e, direction_l, daction_l);
       }
